Исрпавил ошибку парсинга date в String(дни отображались как Err) Добавил Selector для дней и отображение выбранного дня
