+++ app/src/main/java/com/example/bookingapp/domain/entities/Day.kt
@@ -3,4 +3,5 @@ package com.example.bookingapp.domain.entities
 data class Day(
     val id: Int,
     val date: Long,
+    var isSelected: Boolean = false,
 )
