+++ app/src/main/res/drawable/day_clicked.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid
+        android:color="?colorSecondary">
+    </solid>
+
+    <stroke android:width="1dp"
+        android:color="#B9B6B6" />
+
+    <corners android:radius="15dp" />
+</shape>
