+++ app/src/main/res/drawable/day_selector.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:drawable="@drawable/day_clicked"
+        android:state_selected="true">
+    </item>
+    <item
+        android:drawable="@drawable/day_shape">
+    </item>
+</selector>
