+++ app/src/main/res/layout/card_date.xml
@@ -1,19 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    android:gravity="center"
-    android:orientation="vertical"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:paddingVertical="?listPreferredItemPaddingVertical"
     android:paddingStart="?listPreferredItemPaddingStart"
-    android:paddingEnd="?listPreferredItemPaddingEnd">
+    android:paddingEnd="?listPreferredItemPaddingEnd"
+    android:background="@drawable/day_selector"
+    android:layout_margin="2dp">
 
     <TextView
         android:id="@+id/date"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAppearance="?textAppearanceHeadline5"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
         tools:text="8" />
 
     <TextView
@@ -21,6 +26,10 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAppearance="?textAppearanceHeadline6"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/date"
         tools:text="ЧТ" />
 
-</LinearLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
+
