+++ app/src/main/res/layout/companies_item.xml
@@ -13,7 +13,6 @@
         android:paddingStart="?listPreferredItemPaddingStart"
         android:paddingEnd="?listPreferredItemPaddingEnd">
 
-        <!-- TODO #37 -->
         <ImageView
             android:id="@+id/company_logo"
             android:layout_width="30dp"
